Avoid memory allocation in AudioWorkletProcessor.process()
Based on the spec [1], the current implementation of
AudioWorkletProcessor needs to create a new data container
(i.e. WebIDL sequence<>) for input, output, and param arrays.

With the new spec change [2], this CL changes the overall design
of the audio processing callback:

1. Moves the processing call from AudioWorkletGlobalScope to
   AudioWorkletProcessor object.
2. AudioWorkletProcessor now keeps the data container within
   the object and allocate memory when it is needed.

The preliminary benchmark shows the sizable improvement in the
audio stream quality. The glitch score
(= buffer underrun/total callback) is improved by ~9x in the
low-tier machine.

[1]: https://webaudio.github.io/web-audio-api/#dom-audioworkletprocessor-process-inputs-outputs-parameters-inputs
[2]: https://github.com/WebAudio/web-audio-api/issues/1933#issuecomment-616632754
[3]: https://bugs.chromium.org/p/chromium/issues/detail?id=1086665#c2

Bug: 1086665
Change-Id: I3e664754973d4d86649d38c1807c6b9d7830fb96